Refactor: Rename ~cache to ~apollo, move everything apollo related
