+++ src/apollo/apolloClient.ts
@@ -1,6 +1,6 @@
 import { ApolloClient, HttpLink } from '@apollo/client';
 
-import cache from '~cache';
+import cache from './cache';
 
 const httpLink = new HttpLink({
   uri: `${process.env.AUTH_PROXY_ENDPOINT || 'http://localhost:3005'}/graphql`,
