+++ src/apollo/index.ts
@@ -1,0 +1,1 @@
+export { default as apolloClient } from './apolloClient';
