+++ tsconfig.base.json
@@ -29,8 +29,8 @@
       "~utils/*": ["src/utils/*"],
       "~routes": ["src/routes"],
       "~routes/*": ["src/routes/*"],
-      "~cache": ["src/cache"],
-      "~cache/*": ["src/cache/*"],
+      "~apollo": ["src/apollo"],
+      "~apollo/*": ["src/apollo/*"],
       "~transformers": ["src/transformers"],
       "~transformers/*": ["src/transformers/*"],
       "~auth": ["src/auth"],
